Move files in the student original project to the "student" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
