+++ projects-parent/originals-parent/student/pom.xml
@@ -36,7 +36,7 @@
         <configuration>
           <archive>
             <manifest>
-              <mainClass>edu.pdx.cs410J.original.Student</mainClass>
+              <mainClass>edu.pdx.cs410J.student.Student</mainClass>
             </manifest>
           </archive>
         </configuration>
