+++ projects-parent/originals-parent/student/src/it/java/edu/pdx/cs410J/student/StudentIT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.student;
 
 import edu.pdx.cs410J.InvokeMainTestCase;
 import org.junit.Test;
