+++ projects-parent/originals-parent/student/src/test/java/edu/pdx/cs410J/student/StudentTest.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.student;
 
 import org.junit.Test;
